finishing employee

+++ app/Http/Controllers/ProjectController.php
@@ -5,13 +5,23 @@ namespace App\Http\Controllers;
 use App\Models\Project;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use App\Models\Service;
 
 class ProjectController extends Controller
 {
     public function index()
     {
-        $projects = Project::all();
-        return view('dashboard.projects.index', compact('projects'));
+
+        if (auth()->user()->role == 'admin') {
+            $projects = Project::all();
+        } else {
+            $projects = Project::where('employee_id', auth()->id())->get();
+        }
+    
+        // $projects = Project::all();
+        $services = Service::all();
+
+        return view('dashboard.projects.index', compact('projects','services'));
     }
 
     public function show($id)
@@ -22,7 +32,9 @@ class ProjectController extends Controller
 
     public function create()
     {
-        return view('dashboard.projects.create');
+        $services = Service::all();
+
+        return view('dashboard.projects.create', compact('services'));
     }
 
 
@@ -35,6 +47,8 @@ class ProjectController extends Controller
             'content' => 'required|string',
             'images' => 'required',
             'images.*' => 'image',
+            'service_id' => 'required|exists:services,id',
+
         ]);
 
         $imagePath = $request->file('image')->store('projects', 'public');
@@ -46,26 +60,55 @@ class ProjectController extends Controller
             }
         }
 
+        $userId = auth()->id();
+
         Project::create([
             'title' => $validated['title'],
             'image' => $imagePath,
             'description' => $validated['description'],
             'content' => $validated['content'],
             'images' => json_encode($imagesPaths),
+            'service_id' => $validated['service_id'],
+            'employee_id' => $userId, 
+
         ]);
 
-        return redirect()->route('projects.index')->with('success', 'أضاف المشروع بنجاح!');
+        $user = auth()->user();
+
+        if ($user->role === 'admin') {
+            return redirect()->route('projects.index')->with('success', 'أضاف المشروع بنجاح!');
+        }
+        elseif ($user->role === 'employee')
+        {
+            return redirect()->route('employee.projects.index')->with('success', 'أضاف المشروع بنجاح!');
+        }
     }
     public function edit($id)
     {
+
+        $user = auth()->user();
+        $userId = auth()->id();
+        
         $project = Project::findOrFail($id);
-        return view('dashboard.projects.edit', compact('project'));
+
+        if ($user->role == 'employee' && $project->employee_id != $userId) {
+            return abort(403, 'ليس لديك صلاحية التعديل على هذا المشروع.');
+        }
+
+        $services = Service::all(); 
+        return view('dashboard.projects.edit', compact('project','services'));
     }
 
 
     public function update(Request $request, $id)
     {
+        $user = auth()->user();
         $project = Project::findOrFail($id);
+
+        if ($user->role === 'employee' && $project->employee_id !== $user->id) {
+            return abort(403, 'ليس لديك صلاحية التعديل على هذا المشروع.');
+        }
+        
 
         $validated = $request->validate([
             'title' => 'required|string|max:255',
@@ -74,6 +117,7 @@ class ProjectController extends Controller
             'content' => 'required|string',
             'images' => 'nullable',
             'images.*' => 'image',
+            'service_id' => 'required|exists:services,id',
         ]);
 
         if ($request->hasFile('image')) {
@@ -101,15 +145,36 @@ class ProjectController extends Controller
             'title' => $validated['title'],
             'description' => $validated['description'],
             'content' => $validated['content'],
+            'service_id' => $validated['service_id']
+
         ]);
 
-        return redirect()->route('projects.index')->with('success', 'تم تحديث المشروع بنجاح!');
+        if ($user->role === 'admin') {
+            return redirect()->route('projects.index')->with('success', 'تم تحديث المشروع بنجاح!');
+        }
+        elseif ($user->role === 'employee')
+        {
+            return redirect()->route('employee.projects.index')->with('success', 'تم تحديث المشروع بنجاح!');
+        }
+
     }
 
     public function destroy($id)
     {
+        $user = auth()->user();
         $project = Project::findOrFail($id);
+
+        if ($user->role === 'employee' && $project->employee_id !== $user->id) {
+            return abort(403, 'ليس لديك صلاحية حذف هذا المشروع.');
+        }
         $project->delete();
-        return redirect()->route('projects.index')->with('success', 'تم حذف المشروع بنجاح!');
+
+        if ($user->role === 'admin') {
+            return redirect()->route('projects.index')->with('success', 'تم حذف المشروع بنجاح!');
+        }
+        elseif ($user->role === 'employee')
+        {
+            return redirect()->route('employee.projects.index')->with('success', 'تم حذف المشروع بنجاح!');
+        }
     }
 }
